+++ reader/src/ASTR/settings.vue
@@ -1,7 +1,7 @@
 <template>
     <div id="setting">
         <div class="warmtip" style="background-color: rgb(255, 255, 0);color: #000; padding: 5px;" v-show="!hidetip">
-            <span style="float: left;margin-left: 20px;">If you like this project, please give me a star <a href="https://github.com/050644zf/ArknightsStoryTextReader" style="color: rgb(32, 32, 184);">here</a>!</span>
+            <span style="float: left;margin-left: 20px;">If you like this project, please give me a star <a href="https://github.com/050644zf/ArknightsStoryTextReader" style="color: rgb(32, 32, 184);">here</a>!<br/>If you have any issues or feedbacks please summit them <a href="https://github.com/050644zf/ArknightsStoryTextReader/issues" style="color: rgb(32, 32, 184);">here</a>.</span>
             <span class="material-icons" style="float: right;" @click="hide()">clear</span>
             <div style="clear: both;"></div>
         </div>
@@ -11,8 +11,10 @@
             </span>
             <span class="settingtitle">{{i18n.setting[lang]}}</span>
             
-            <span class="material-icons clearbutton" style="float: right;padding: 4px;" @click="clearSetting()">delete</span>
-            <span class="settingtitle" style="float: right; font-weight: normal;font-size: 15px;">{{i18n.clear[lang]}}: </span>
+            <div class="clear" @click="clearSetting()">
+                <span class="material-icons clearbutton" style="float: right;padding: 4px;" >delete</span>
+                <span class="settingtitle" style="float: right; font-weight: normal;font-size: 15px;">{{i18n.clear[lang]}} </span>                
+            </div>
             <div style="clear: both;"></div>
             <div :class="{settingoptions:true}" v-show="show">
                 <span style="float: left;">{{i18n.dr[lang] }}: </span>
@@ -24,9 +26,12 @@
                 <span style="float: left;">{{i18n.hideName[lang]}}(beta): </span>
                 <input class="checkbox" type="checkbox" v-model="hideName" true-value="y" false-value="n" style="float:left"/>
                 <div style="clear:both;"></div>
-                <span style="float: left; font-weight: bold;">{{i18n.save[lang] }}: </span>
-                <span class="material-icons" @click="save()" style="float: left;">save_alt</span>
-                <div style="clear: both;"></div>
+                <div class="save" @click="save()">
+                    <span style="float: left; font-weight: bold;">{{i18n.save[lang] }} </span>
+                    <span class="material-icons"  style="float: left;">save_alt</span>
+                    <div style="clear: both;"></div>
+                </div>
+
                 <span style="font-size: 10px; font-weight: bold; color: dimgray;">
                     Developer Info/Options (Please attach these info with your issues)
                 </span>
@@ -129,6 +134,19 @@ methods:{
 .settingshow .settingtitle{
     color:rgba(255, 255, 255, 1);
 }
+
+.clear{
+    background-color: #FF000000;
+    margin: 4px;
+    padding: 0px 5px;
+    width: fit-content;
+    transition: background-color 0.5s;
+    float: right;
+}
+.clear:hover{
+    background-color: #FF0000FF;
+}
+
 .clearbutton{
     color: rgba(255, 255, 255, 0);
     transition: color 0.5s;
@@ -159,6 +177,17 @@ methods:{
     width:20px;
     height: 20px;
     margin: 4px;
+}
+
+.save{
+    width: fit-content;
+    margin: 4px;
+    padding: 0px 5px;
+    background-color: #00808088;
+    transition: background-color 0.5s;
+}
+.save:hover{
+    background-color: #008080FF;
 }
 
 </style>